refactor: small tweaks
