+++ app/frontend/entrypoints/application.js
@@ -36,7 +36,7 @@ import "/js/TestTransformation";
 const tooltipTriggerList = document.querySelectorAll(
   '[data-bs-toggle="tooltip"]'
 );
-const tooltipList = [...tooltipTriggerList].map(
+[...tooltipTriggerList].map(
   (tooltipTriggerEl) => new bootstrap.Tooltip(tooltipTriggerEl)
 );
