+++ app/frontend/js/react.jsx
@@ -3,7 +3,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 
-import TransformationApp from "~/js/apps/TransformationApp";
+import TransformationApp from "~/js/apps/TransformationApp/TransformationApp";
 
 import { Provider } from "react-redux";
 import configureAppStore from "~/js/store";
