+++ spec/javascript/TransformApp/components/index.spec.js
@@ -1,12 +1,10 @@
 import React from 'react'
 import { fireEvent, screen } from '@testing-library/react'
-import TransformationApp from 'js/apps/TransformationApp/index';
+import TransformationApp from 'js/apps/TransformationApp/TransformationApp';
 import { renderWithProviders } from '../../test-utils';
 
 describe("<TransformationApp />", () => {
   it("runs", () => {
     renderWithProviders(<TransformationApp />);
-
-    
   });
 });
